Fix bug that created wrong estimates when there is no delay
When users don't specify delay in the algorithm, it is assumed that
no delay will be applied in the signal. In that case, the default
value for that was set to 0, which made the slice evaluation return
empty always since no delay technically means delay 1 instead of 0.
The default value for the fit function now has changed to 1 to reclect
this logic.